Version 1.1.3
Changes in the arithmetic dealing with the conversion of 'unix' time units to 'ntp' time units has improved the stability of NetAssociation's returned times.  NTP's finest time resolution is 233 picoSeconds and Darwin's is 1 microSecond so I think the conversion was very sensitive when dealing with small time differences.  Tests in the iOS simulator show all the NetAssociations returning similar time offsets with small standard deviations.  There are some more changes in the works to isolate NetAssociations from the NetClock but this release should be, at least, no worse than the previous one.